fix: use OneDrive client abstraction for file operations
