+++ .env.example
@@ -9,6 +9,8 @@ GOOGLE_CLIENT_SECRET=
 MICROSOFT_CLIENT_ID=
 MICROSOFT_CLIENT_SECRET=
 
+ONEDRIVE_ACCESS_TOKEN=
+
 # To generate a secret, just run `openssl rand -base64 32`
 BETTER_AUTH_SECRET=
 BETTER_AUTH_URL=http://localhost:1284
