+++ apps/server/src/providers/onedrive/one-drive.ts
@@ -1,4 +1,4 @@
-import { OneDrive } from "@nimbus/server/lib/one-drive/src/index";
+import { OneDrive } from "@nimbus/server/lib/one-drive/src/client";
 import type { File } from "@/providers/onedrive/types";
 
 export class OneDriveProvider {
@@ -14,17 +14,15 @@ export class OneDriveProvider {
 	async createFile(name: string, mimeType: string, parents?: string[]): Promise<File | null> {
 		const parentId = parents?.[0] || "root";
 
-		const res = await fetch(`https://graph.microsoft.com/v1.0/me/drive/items/${parentId}/children`, {
-			method: "POST",
-			headers: {
-				Authorization: `Bearer ${this.accessToken}`,
-				"Content-Type": "application/json",
-			},
-			body: JSON.stringify({
+		const res: Response = await this.drive.post(`/me/drive/items/${parentId}/children`, {
+			body: {
 				name,
 				folder: mimeType === "application/vnd.microsoft.folder" ? {} : undefined,
 				file: mimeType !== "application/vnd.microsoft.folder" ? {} : undefined,
-			}),
+			},
+			headers: {
+				Authorization: `Bearer ${this.accessToken}`,
+			},
 		});
 
 		if (!res.ok) return null;
@@ -32,20 +30,19 @@ export class OneDriveProvider {
 	}
 
 	async listFiles(): Promise<File[]> {
-		const res = await fetch("https://graph.microsoft.com/v1.0/me/drive/root/children", {
+		const res: Response = await this.drive.get(`/me/drive/root/children`, {
 			headers: {
 				Authorization: `Bearer ${this.accessToken}`,
 			},
 		});
 
 		if (!res.ok) return [];
-
 		const data = (await res.json()) as { value: File[] };
-		return data.value as File[];
+		return data.value;
 	}
 
 	async getFileById(id: string): Promise<File | null> {
-		const res = await fetch(`https://graph.microsoft.com/v1.0/me/drive/items/${id}`, {
+		const res: Response = await this.drive.get(`/me/drive/items/${id}`, {
 			headers: {
 				Authorization: `Bearer ${this.accessToken}`,
 			},
@@ -56,13 +53,12 @@ export class OneDriveProvider {
 	}
 
 	async updateFile(fileId: string, name: string): Promise<File | null> {
-		const res = await fetch(`https://graph.microsoft.com/v1.0/me/drive/items/${fileId}`, {
-			method: "PATCH",
+		const res: Response = await this.drive.patch(`/me/drive/items/${fileId}`, {
+			body: { name },
 			headers: {
 				Authorization: `Bearer ${this.accessToken}`,
 				"Content-Type": "application/json",
 			},
-			body: JSON.stringify({ name }),
 		});
 
 		if (!res.ok) return null;
@@ -70,8 +66,7 @@ export class OneDriveProvider {
 	}
 
 	async deleteFile(fileId: string): Promise<boolean> {
-		const res = await fetch(`https://graph.microsoft.com/v1.0/me/drive/items/${fileId}`, {
-			method: "DELETE",
+		const res: Response = await this.drive.delete(`/me/drive/items/${fileId}`, {
 			headers: {
 				Authorization: `Bearer ${this.accessToken}`,
 			},
